Considering dutyTypes for events creating
